Workaround for a small issue
Apparently, it seems that the search is not working when the filenames are of the type: "A Quite Place", "X men", etc.
These are file names that start with a single letter A, a, X, x, etc (but not Q, q, Z, z, etc) followed by a space. In the above mentioned patch, If this type of pattern is found and the starting letter in the list that causes the error, then we remove the first two characters from the file name.

(This is just a workaround) because the actual code itself is correct. It seems that the google drive library itself is returning an empty list for these types of fileNames.